Move prediction pipeliens to separate folder reduce clutter
