+++ pipelines/__init__.py
@@ -24,3 +24,4 @@ from .other_models import *
 from .selftrain import *
 from .spl import *
 from .training import *
+from .predicting import *
